novartis_datathon_2025-Arman/requirements.txt
# Novartis Datathon 2025 - Pinned Requirements
# Generated for reproducibility. Tested with Python 3.10+
# Install: pip install -r requirements.txt

# Core ML/DL frameworks
torch>=2.6.0
torchvision>=0.17.0

# Data processing
numpy==1.26.4
pandas==2.3.3
pyarrow==22.0.0

# Visualization
matplotlib==3.10.7
seaborn==0.13.2
plotly==6.5.0

# Gradient boosting models
lightgbm==4.6.0
xgboost==3.1.2
catboost==1.2.8

# Machine learning utilities
scikit-learn==1.7.2
joblib==1.5.2

# Time series models
statsmodels==0.14.4

# Optimization
scipy==1.15.3

# Deep learning utilities
tqdm==4.67.1

# Jupyter/Colab
ipykernel==7.1.0
ipywidgets==8.1.8

# Configuration
pyyaml==6.0.3
omegaconf==2.3.0

# Testing
pytest==9.0.1

# Experiment tracking (optional - uncomment to install)
# wandb>=0.16.0
# mlflow>=2.10.0

# Hyperparameter optimization (optional - uncomment to install)
# optuna>=3.5.0
